+++ src/hooks/validators/getStringValidator.ts
@@ -5,33 +5,27 @@ import { ErrorTypes } from './types'
 
 export const getStringValidator = (
   currentObject: JSONSchemaType,
-  required: boolean
+  baseValidator: ValidationOptions
 ): ValidationOptions => {
-  const validator: ValidationOptions = {}
-
-  if (required) {
-    validator.required = ErrorTypes.required
-  }
-
   if (currentObject.minLength) {
-    validator.minLength = {
+    baseValidator.minLength = {
       value: currentObject.minLength,
       message: ErrorTypes.minLength,
     }
   }
 
   if (currentObject.maxLength) {
-    validator.maxLength = {
+    baseValidator.maxLength = {
       value: currentObject.maxLength,
       message: ErrorTypes.maxLength,
     }
   }
 
   if (currentObject.pattern) {
-    validator.pattern = {
+    baseValidator.pattern = {
       value: new RegExp(currentObject.pattern),
       message: ErrorTypes.pattern,
     }
   }
-  return validator
+  return baseValidator
 }