+++ src/hooks/validators/index.ts
@@ -1,5 +1,5 @@
-export * from './getBooleanValidator'
-export * from './getNumberValidator'
+export * from './numberUtilities'
 export * from './getStringValidator'
+export * from './getGenericValidator'
 export * from './getError'
 export * from './types'